+++ src/com/iispl/service/TransactionService.java
@@ -1,6 +1,7 @@
 package com.iispl.service;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
+
 
 import com.iispl.entity.Transaction;
 
@@ -8,6 +9,6 @@ public interface TransactionService {
 
     public void addTransaction(Transaction transaction);
     public void displayAllTransactions();
-    public void displayHighValueTransactions(BigInteger threshold);
+    public void displayHighValueTransactions(BigDecimal threshold);
     public void displayFailedTransactions();
 }
